+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="cz.muni.fi.pv256.movio2.uco_410371.MoviesActivity"
+            android:name="cz.muni.fi.pv256.movio2.uco_410371.movies.MoviesActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme">
             <intent-filter>
@@ -37,10 +37,10 @@
                 <!--android:name="android.support.PARENT_ACTIVITY"-->
                 <!--android:value=".MoviesActivity"/>-->
         <!--</activity>-->
-        <activity android:name=".MoviesPreferences">
+        <activity android:name=".movies.MoviesPreferences">
         </activity>
         <service
-            android:name=".network.DownloadService"
+            android:name=".DownloadService"
             android:enabled="true"
             android:exported="true"/>
         <provider
