+++ app/src/main/java/cz/muni/fi/pv256/movio2/uco_410371/ItemClickListener.java
@@ -1,4 +1,4 @@
-package cz.muni.fi.pv256.movio2.uco_410371.interfaces;
+package cz.muni.fi.pv256.movio2.uco_410371;
 
 import android.view.View;
