+++ app/src/main/java/cz/muni/fi/pv256/movio2/uco_410371/movies/MoviesActivity.java
@@ -1,7 +1,5 @@
-package cz.muni.fi.pv256.movio2.uco_410371;
+package cz.muni.fi.pv256.movio2.uco_410371.movies;
 
-import android.content.Context;
-import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
@@ -10,6 +8,7 @@ import android.view.MenuItem;
 import android.widget.CompoundButton;
 import android.widget.Switch;
 
+import cz.muni.fi.pv256.movio2.uco_410371.R;
 import cz.muni.fi.pv256.movio2.uco_410371.sync.UpdaterSyncAdapter;
 import cz.muni.fi.pv256.movio2.uco_410371.util.ActivityUtils;
