+++ app/src/main/java/cz/muni/fi/pv256/movio2/uco_410371/movies/MoviesFragment.java
@@ -1,39 +1,41 @@
-package cz.muni.fi.pv256.movio2.uco_410371;
+package cz.muni.fi.pv256.movio2.uco_410371.movies;
 
 import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
+import android.support.v4.content.LocalBroadcastManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
 
-import java.util.ArrayList;
-import java.util.List;
+import cz.muni.fi.pv256.movio2.uco_410371.R;
+import cz.muni.fi.pv256.movio2.uco_410371.adapters.HorizontalRecyclerViewAdapter;
 
-import cz.muni.fi.pv256.movio2.uco_410371.adapters.MovieVERTRecyclerViewAdapter;
-import cz.muni.fi.pv256.movio2.uco_410371.db.MovioManager;
-import cz.muni.fi.pv256.movio2.uco_410371.db.models.MovieTable;
-import cz.muni.fi.pv256.movio2.uco_410371.models.Movie;
-
-public class MoviesDBFragment extends Fragment {
+public class MoviesFragment extends Fragment implements MoviesContract.View {
 
     //*****CONSTANT*****
-    public static final String TAG = MoviesDBFragment.class.getName();
+    public static final String MESSAGE = "message";
+    public static final String TAG = MoviesFragment.class.getName();
 
     private boolean mTwoPane;
+    private HorizontalRecyclerViewAdapter mRecyclerViewAdapter;
 
-    public MoviesDBFragment() {}
+    private MoviesContract.Presenter mPresenter;
+
+    public MoviesFragment() {}
+
+    public static MoviesFragment newInstance() {
+        return new MoviesFragment();
+    }
 
     @Override
     public void onAttach(Context context) {
         super.onAttach(context);
         Log.d(TAG, "onAttach: ");
-        Toast.makeText(context, "DB fragment", Toast.LENGTH_SHORT).show();
     }
 
     @Override
@@ -41,6 +43,8 @@ public class MoviesDBFragment extends Fragment {
         super.onCreate(savedInstanceState);
         Log.d(TAG, "onCreate: ");
         setRetainInstance(true);
+        mPresenter = new MoviesPresenter(this, getActivity(), LocalBroadcastManager.getInstance(getActivity()));
+        mPresenter.startService();
     }
 
     @Nullable
@@ -57,10 +61,8 @@ public class MoviesDBFragment extends Fragment {
         recyclerView.setHasFixedSize(true);
         recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
 
-        List<MovieTable> movieTables = new MovioManager(getContext()).getAllMovies();
-
-        MovieVERTRecyclerViewAdapter movieVERTRecyclerViewAdapter = new MovieVERTRecyclerViewAdapter(getContext(), convertMovieTableToMovie(movieTables), mTwoPane);
-        recyclerView.setAdapter(movieVERTRecyclerViewAdapter);
+        mRecyclerViewAdapter = new HorizontalRecyclerViewAdapter(getContext(), mTwoPane);
+        recyclerView.setAdapter(mRecyclerViewAdapter);
 
         return view;
     }
@@ -68,25 +70,27 @@ public class MoviesDBFragment extends Fragment {
     @Override
     public void onStart() {
         super.onStart();
-        Log.d(MoviesDBFragment.TAG, "onStart: ");
+        Log.d(TAG, "onStart: ");
     }
 
     @Override
     public void onResume() {
         super.onResume();
-        Log.d(MoviesDBFragment.TAG, "onResume: ");
+        Log.d(TAG, "onResume: ");
+        mPresenter.registerReceiver(mRecyclerViewAdapter);
     }
 
     @Override
     public void onPause() {
         super.onPause();
-        Log.d(MoviesDBFragment.TAG, "onPause: ");
+        Log.d(TAG, "onPause: ");
+        mPresenter.unregisterReceiver();
     }
 
     @Override
     public void onStop() {
         super.onStop();
-        Log.d(MoviesDBFragment.TAG, "onStop: ");
+        Log.d(TAG, "onStop: ");
     }
 
     @Override
@@ -105,14 +109,5 @@ public class MoviesDBFragment extends Fragment {
     public void onDetach() {
         super.onDetach();
         Log.d(TAG, "onDetach: ");
-    }
-
-    private List<Movie> convertMovieTableToMovie(List<MovieTable> movieTables) {
-        List<Movie> converted = new ArrayList<>();
-        for (MovieTable mt : movieTables) {
-            converted.add(new Movie(mt.getTMDId(), mt.getReleaseDate(), mt.getPosterPath(),
-                    mt.getTitle(), mt.getBackdropPath(), (float)mt.getPopularity()));
-        }
-        return converted;
     }
 }
