+++ app/src/main/java/cz/muni/fi/pv256/movio2/uco_410371/movies/MoviesPreferences.java
@@ -1,10 +1,12 @@
-package cz.muni.fi.pv256.movio2.uco_410371;
+package cz.muni.fi.pv256.movio2.uco_410371.movies;
 
 import android.app.FragmentManager;
 import android.app.FragmentTransaction;
 import android.preference.PreferenceFragment;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import cz.muni.fi.pv256.movio2.uco_410371.R;
 
 public class MoviesPreferences extends AppCompatActivity {
