+++ build.gradle
@@ -22,3 +22,21 @@ allprojects {
 task clean(type: Delete) {
     delete rootProject.buildDir
 }
+
+ext {
+    // SDK and Tools
+    minSdkVersion = 17
+    targetSdkVersion = 24
+    compileSdkVersion = 24
+    buildToolsVersion = '24.0.3'
+
+    // Dependencies version
+    supportLibraryVersion = '24.2.1'
+    guavaLibraryVersion = '18.0'
+    picassoLibraryVersion = '2.5.2'
+    logansquareLibraryVersion = '1.3.6'
+    logansquareConverterLibraryVersion = '1.4.1'
+    retrofit2LibraryVersion = '2.1.0'
+    mosbyLibraryVersion = '2.0.1'
+
+}
